int -> number(1)

+++ conf/evolutions/default/70.sql
@@ -1,5 +1,5 @@
 # --- !Ups
-alter table USER_PREFERENCE add CHOSEN_COLOUR_SCHEME int default 1 not null;
+alter table USER_PREFERENCE add CHOSEN_COLOUR_SCHEME number(1) default 1 not null;
 
 # --- !Downs
 alter table USER_PREFERENCE drop column CHOSEN_COLOUR_SCHEME;
